# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campaign Nexus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns/{campaign_id}/clock": {
      "get": {
        "tags": [
          "time"
        ],
        "summary": "Get Clock",
        "operationId": "get_clock_api_v1_campaigns__campaign_id__clock_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/clock/realtime": {
      "post": {
        "tags": [
          "time"
        ],
        "summary": "Set Realtime",
        "operationId": "set_realtime_api_v1_campaigns__campaign_id__clock_realtime_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealtimeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/clock/advance": {
      "post": {
        "tags": [
          "time"
        ],
        "summary": "Advance",
        "operationId": "advance_api_v1_campaigns__campaign_id__clock_advance_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/clock/advance/preview": {
      "post": {
        "tags": [
          "time"
        ],
        "summary": "Advance Preview",
        "operationId": "advance_preview_api_v1_campaigns__campaign_id__clock_advance_preview_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/scheduled-events": {
      "get": {
        "tags": [
          "time"
        ],
        "summary": "List Scheduled",
        "operationId": "list_scheduled_api_v1_campaigns__campaign_id__scheduled_events_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduledEventOut"
                  },
                  "title": "Response List Scheduled Api V1 Campaigns  Campaign Id  Scheduled Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "time"
        ],
        "summary": "Create Scheduled",
        "operationId": "create_scheduled_api_v1_campaigns__campaign_id__scheduled_events_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledEventCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/scheduled-events/{event_id}": {
      "delete": {
        "tags": [
          "time"
        ],
        "summary": "Cancel Scheduled",
        "operationId": "cancel_scheduled_api_v1_campaigns__campaign_id__scheduled_events__event_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/flags": {
      "get": {
        "tags": [
          "time"
        ],
        "summary": "Get Flags",
        "operationId": "get_flags_api_v1_campaigns__campaign_id__flags_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Flags Api V1 Campaigns  Campaign Id  Flags Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdvancePreview": {
        "properties": {
          "from_time": {
            "type": "integer",
            "title": "From Time"
          },
          "to_time": {
            "type": "integer",
            "title": "To Time"
          },
          "would_fire": {
            "items": {
              "$ref": "#/components/schemas/FiredEvent"
            },
            "type": "array",
            "title": "Would Fire",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "from_time",
          "to_time"
        ],
        "title": "AdvancePreview"
      },
      "AdvanceReport": {
        "properties": {
          "from_time": {
            "type": "integer",
            "title": "From Time"
          },
          "to_time": {
            "type": "integer",
            "title": "To Time"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "formatted": {
            "$ref": "#/components/schemas/ClockFormatted"
          },
          "fired": {
            "items": {
              "$ref": "#/components/schemas/FiredEvent"
            },
            "type": "array",
            "title": "Fired",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "from_time",
          "to_time",
          "reason",
          "formatted"
        ],
        "title": "AdvanceReport"
      },
      "AdvanceRequest": {
        "properties": {
          "days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Days",
            "default": 0
          },
          "hours": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Hours",
            "default": 0
          },
          "minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Minutes",
            "default": 0
          },
          "seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seconds",
            "default": 0
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "manual"
          }
        },
        "type": "object",
        "title": "AdvanceRequest"
      },
...
      "ClockFormatted": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "weekday": {
            "type": "string",
            "title": "Weekday"
          },
          "month": {
            "type": "string",
            "title": "Month"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "seconds": {
            "type": "integer",
            "title": "Seconds"
          }
        },
        "type": "object",
        "required": [
          "label",
          "weekday",
          "month",
          "day",
          "year",
          "time",
          "season",
          "seconds"
        ],
        "title": "ClockFormatted"
      },
      "ClockOut": {
        "properties": {
          "time_game": {
            "type": "integer",
            "title": "Time Game"
          },
          "calendar_name": {
            "type": "string",
            "title": "Calendar Name"
          },
          "calendar": {
            "additionalProperties": true,
            "type": "object",
            "title": "Calendar"
          },
          "realtime_enabled": {
            "type": "boolean",
            "title": "Realtime Enabled"
          },
          "realtime_paused": {
            "type": "boolean",
            "title": "Realtime Paused"
          },
          "formatted": {
            "$ref": "#/components/schemas/ClockFormatted"
          }
        },
        "type": "object",
        "required": [
          "time_game",
          "calendar_name",
          "calendar",
          "realtime_enabled",
          "realtime_paused",
          "formatted"
        ],
        "title": "ClockOut"
      },
...
      "FiredEvent": {
        "properties": {
          "scheduled_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Event Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "at_time": {
            "type": "integer",
            "title": "At Time"
          },
          "at_label": {
            "type": "string",
            "title": "At Label"
          },
          "narrative": {
            "type": "string",
            "title": "Narrative"
          }
        },
        "type": "object",
        "required": [
          "scheduled_event_id",
          "title",
          "at_time",
          "at_label",
          "narrative"
        ],
        "title": "FiredEvent"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RealtimeRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "RealtimeRequest"
      },
...
      "ScheduledEventCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "fire_at_game": {
            "type": "integer",
            "title": "Fire At Game"
          },
          "action_type": {
            "type": "string",
            "pattern": "^(narrate|set_flag)$",
            "title": "Action Type"
          },
          "action_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action Json",
            "default": {}
          },
          "recurrence_days": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Days"
          }
        },
        "type": "object",
        "required": [
          "title",
          "fire_at_game",
          "action_type"
        ],
        "title": "ScheduledEventCreate"
      },
      "ScheduledEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "fire_at_game": {
            "type": "integer",
            "title": "Fire At Game"
          },
          "fire_at_label": {
            "type": "string",
            "title": "Fire At Label"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "recurrence_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Days"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "fire_at_game",
          "fire_at_label",
          "action_type",
          "recurrence_days",
          "status"
        ],
        "title": "ScheduledEventOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}